+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- The `--border` flag can be used to create a border around the ascii image
+- Two more tests, which fully compare the results
+
 ## [0.5.1] - 2022-03-14
 
 ### Changed